requirements.txt
streamlit>=1.30.0
earthengine-api>=0.1.335
geemap>=0.24.2
folium>=0.17.1
numpy>=1.26.0
pandas>=2.1.0
google-auth>=2.22.0
plotly>=5.20.0